+++ datamodules/dataset.py
@@ -11,52 +11,30 @@ class NarrativeDataset(Dataset):
     def __init__(
         self,
         split,
-        size_dataset,
         path_data,
         lc,
         nc,
-        n_shards,
     ):
 
         self.split = split
-        self.n_shards = n_shards
-        self.size_dataset = size_dataset
         self.lc = lc
         self.nc = nc
-        self.curent_ith_file = -1
 
-        path_data = path_data.replace("[SPLIT]", split).replace("[SHARD]", "*")
-        self.paths = sorted(glob.glob(path_data))
-
-        self.q_ids = None
-        self.c_ids = None
-        self.a1_ids = None
-        self.a2_ids = None
+        self.q_ids = []
+        self.c_ids = []
+        self.a1_ids = []
+        self.a2_ids = []
 
         self.exchange_rate = 0
 
+        self.read_datasetfile(path_data, split)
+
     def __len__(self) -> int:
-        return self.size_dataset
+        return len(self.q_ids)
 
     def __getitem__(self, indx):
         if torch.is_tensor(indx):
             indx = indx.tolist()
-
-        size_shard = self.size_dataset // self.n_shards
-
-        ith_file = indx // size_shard
-        indx = indx % size_shard
-
-        if ith_file == self.n_shards:
-            ith_file -= 1
-            indx = indx + size_shard
-
-        # Check nth file and reload dataset if needed
-        if ith_file != self.curent_ith_file:
-            self.curent_ith_file = ith_file
-
-            # Reload dataset
-            self.read_datasetfile(self.paths[self.curent_ith_file])
 
         return {
             "q_ids": self.q_ids[indx],
@@ -65,37 +43,16 @@ class NarrativeDataset(Dataset):
             "a2_ids": self.a2_ids[indx],
         }
 
-    def _get_context(self, En, Hn):
-        n_samples = min((len(En), self.nc))
-        if self.split == "train":
-            selects_Hn = int(n_samples * self.exchange_rate)
-            selects_En = n_samples - selects_Hn
-
-            return sample(En, selects_En) + sample(Hn, selects_Hn)
-
-        return sample(Hn, n_samples)
-
-    def read_datasetfile(self, path_file):
-        df = pd.read_parquet(path_file)
-
-        self.q_ids = []
-        self.c_ids = []
-        self.a1_ids = []
-        self.a2_ids = []
+    def read_datasetfile(self, path_data, split):
+        df = pd.read_parquet(path_data.replace("[SPLIT]", split))
 
         for entry in df.itertuples():
-            q = entry.q_ids[(entry.q_ids != 101) & (entry.q_ids != 102)]
-            a1 = entry.a1_ids[(entry.a1_ids != 101) & (entry.a1_ids != 102)]
-            a2 = entry.a1_ids[(entry.a1_ids != 101) & (entry.a1_ids != 102)]
-            cE = entry.c_E_ids[(entry.c_E_ids != 101) & (entry.c_E_ids != 102)]
-            cE = np.reshape(cE, (-1, self.lc))
-            cH = entry.c_H_ids[(entry.c_H_ids != 101) & (entry.c_H_ids != 102)]
-            cH = np.reshape(cH, (-1, self.lc))
+            self.q_ids.append(np.copy(entry.q_ids))
+            self.a1_ids.append(np.copy(entry.a1_ids))
+            self.a2_ids.append(np.copy(entry.a2_ids))
 
-            self.q_ids.append(np.copy(q))
-            self.a1_ids.append(np.copy(a1))
-            self.a2_ids.append(np.copy(a2))
-
+            cE = np.reshape(entry.c_E_ids, (-1, self.lc))
+            cH = np.reshape(entry.c_H_ids, (-1, self.lc))
             c = np.zeros((self.nc, self.lc), dtype=int)
             n_samples = cE.shape[0]
             if self.split == "train":